Respect iframe offset and remove outdated collision detection
