+++ assets/js/src/sdk/utils/index.ts
@@ -20,6 +20,7 @@ export * from '@Pimcore/utils/date-time'
 export * from '@Pimcore/utils/files'
 export * from '@Pimcore/utils/helpers'
 export * from '@Pimcore/utils/html'
+export * from '@Pimcore/utils/iframe'
 export * from '@Pimcore/utils/number'
 export * from '@Pimcore/utils/polling-helper'
 export * from '@Pimcore/utils/query-string'
